requirements.txt
requests
schedule
python-telegram-bot
python-dotenv 